add zitify decorator move MonkeyPatch into a separate file
